requirements.txt
Django==4.2.2
djangorestframework==3.14.0
pytils==0.4.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-cors-headers==4.2.0
bcrypt==4.0.1
